fi_api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
load-dotenv==0.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytryfi==0.0.20
requests==2.31.0
sentry-sdk==1.38.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
